+++ guillotina_gcloudstorage/storage.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from datetime import datetime
 from typing import AsyncIterator
 from urllib.parse import quote_plus
 
@@ -33,11 +32,11 @@ from guillotina.schema import Object
 from guillotina.utils import apply_coroutine
 from guillotina.utils import get_authenticated_user_id
 from guillotina.utils import get_current_request
-from guillotina.utils import run_async
 from guillotina.utils import to_str
 from guillotina_gcloudstorage.interfaces import IGCloudBlobStore
 from guillotina_gcloudstorage.interfaces import IGCloudFile
 from guillotina_gcloudstorage.interfaces import IGCloudFileField
+from oauth2client.client import AccessTokenInfo
 from oauth2client.service_account import ServiceAccountCredentials
 
 
@@ -404,9 +403,15 @@ class GCloudBlobStore(object):
             "uniform_bucket_level_access", False
         )
         self._cached_buckets = []
-        self._creation_access_token = datetime.now()
         self._client = None
         self._session = None
+        self._lock = None
+
+    @property
+    def lock(self):
+        if self._lock is None:
+            self._lock = asyncio.Lock()
+        return self._lock
 
     @property
     def session(self):
@@ -414,13 +419,16 @@ class GCloudBlobStore(object):
             self._session = aiohttp.ClientSession()
         return self._session
 
-    def _get_access_token(self):
-        access_token = self._credentials.get_access_token()
-        self._creation_access_token = datetime.now()
-        return access_token.access_token
-
     async def get_access_token(self):
-        return await run_async(self._get_access_token)
+        if self._credentials.access_token is None or self._credentials.access_token_expired:
+            async with self.lock:
+                if self._credentials.access_token is None or self._credentials.access_token_expired:
+                    # re-check after getting lock
+                    loop = self._loop or asyncio.get_event_loop()
+                    return await loop.run_in_executor(None, self._credentials.refresh, None)
+
+        return AccessTokenInfo(
+            access_token=self._credentials.access_token, expires_in=self._credentials._expires_in())
 
     def get_client(self):
         if self._client is None: